added license text to new files

+++ EazeGraphLibrary/src/main/java/org/eazegraph/lib/charts/ValueLineChart.java
@@ -954,7 +954,7 @@ public class ValueLineChart extends BaseChart {
             super.onSizeChanged(w, h, oldw, oldh);
             mGraphWidth = w;
             mGraphHeight = h;
-            mUseableGraphHeight = (int) (mGraphHeight - mTopPadding);
+            mUseableGraphHeight = mGraphHeight - mTopPadding;
         }
 
         /**
